+++ Sources/TagKit/_Deprecations/Deprecated.swift
@@ -1,0 +1,14 @@
+import SwiftUI
+
+@available(*, deprecated, renamed: "TagToggleList")
+public typealias TagEditList = TagToggleList
+
+public extension Tagged {
+
+    @available(*, deprecated, renamed: "slugifiedTags(withConfiguration:)")
+    func slugifiedTags(
+        configuration: SlugConfiguration = .standard
+    ) -> [String] {
+        slugifiedTags(withConfiguration: configuration)
+    }
+}